+++ client/lib/geometry.cpp
@@ -39,10 +39,8 @@ namespace autd {
 class Device {
  public:
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  static std::shared_ptr<Device> Create(int device_id, Eigen::Vector3d position, Eigen::Quaterniond quaternion) {
-    auto device = std::make_shared<Device>();
-
-    device->device_id = device_id;
+  static std::unique_ptr<Device> Create(Eigen::Vector3d position, Eigen::Quaterniond quaternion) {
+    auto device = std::make_unique<Device>();
 
     const Eigen::Affine3d transform_matrix = Eigen::Translation3d(position) * quaternion;
     device->x_direction = quaternion * Eigen::Vector3d(1, 0, 0);
@@ -61,15 +59,14 @@ class Device {
     return device;
   }
 
-  static std::shared_ptr<Device> Create(int device_id, Eigen::Vector3d position, Eigen::Vector3d euler_angles) {
+  static std::unique_ptr<Device> Create(Eigen::Vector3d position, Eigen::Vector3d euler_angles) {
     const auto quaternion = Eigen::AngleAxis(euler_angles.x(), Eigen::Vector3d::UnitZ()) *
                             Eigen::AngleAxis(euler_angles.y(), Eigen::Vector3d::UnitY()) *
                             Eigen::AngleAxis(euler_angles.z(), Eigen::Vector3d::UnitZ());
 
-    return Create(device_id, position, quaternion);
+    return Create(position, quaternion);
   }
 
-  int device_id = 0;
   Eigen::Matrix<double, 3, NUM_TRANS_IN_UNIT> global_trans_positions;
   Eigen::Vector3d x_direction;
   Eigen::Vector3d y_direction;
@@ -85,32 +82,27 @@ class AUTDGeometry : public Geometry {
 
   const int numDevices() noexcept final;
   const int numTransducers() noexcept final;
-  int GroupIDForDeviceID(int device_id) final;
+  int GroupIDForDeviceIdx(int device_id) final;
   const Vector3 position(int transducer_idx) final;
   const Vector3 local_position(int device, Vector3 global_position) final;
   const Vector3 direction(int transducer_id) final;
   const Vector3 x_direction(int transducer_id) final;
   const Vector3 y_direction(int transducer_id) final;
   const Vector3 z_direction(int transducer_id) final;
-  const int deviceIdForTransIdx(int transducer_idx) final;
-  const int deviceIdForDeviceIdx(int device_index) final;
+  const int deviceIdxForTransIdx(int transducer_idx) final;
 
  private:
-  std::vector<std::shared_ptr<Device>> _devices;
+  std::vector<std::unique_ptr<Device>> _devices;
   std::map<int, int> _group_map;
-  std::shared_ptr<Device> device(int transducer_id) {
-    const auto eid = transducer_id / NUM_TRANS_IN_UNIT;
-    return this->_devices.at(eid);
-  }
 };
 
 GeometryPtr Geometry::Create() { return std::make_shared<AUTDGeometry>(); }
 
 int AUTDGeometry::AddDevice(Vector3 position, Vector3 euler_angles, int group) {
-  const auto device_id = static_cast<int>(this->_devices.size());
+  const int device_id = static_cast<int>(this->_devices.size());
   const auto pos = Eigen::Vector3d(position.x(), position.y(), position.z());
   const auto ea = Eigen::Vector3d(euler_angles.x(), euler_angles.y(), euler_angles.z());
-  this->_devices.push_back(Device::Create(device_id, pos, ea));
+  this->_devices.push_back(Device::Create(pos, ea));
   this->_group_map[device_id] = group;
   return device_id;
 }
@@ -119,7 +111,7 @@ int AUTDGeometry::AddDeviceQuaternion(Vector3 position, Quaternion quaternion, i
   const auto device_id = static_cast<int>(this->_devices.size());
   const auto pos = Eigen::Vector3d(position.x(), position.y(), position.z());
   const auto qua = Eigen::Quaterniond(quaternion.w(), quaternion.x(), quaternion.y(), quaternion.z());
-  this->_devices.push_back(Device::Create(device_id, pos, qua));
+  this->_devices.push_back(Device::Create(pos, qua));
   this->_group_map[device_id] = group;
   return device_id;
 }
@@ -128,44 +120,41 @@ const int AUTDGeometry::numDevices() noexcept { return static_cast<int>(this->_d
 
 const int AUTDGeometry::numTransducers() noexcept { return this->numDevices() * NUM_TRANS_IN_UNIT; }
 
-int AUTDGeometry::GroupIDForDeviceID(int device_id) { return this->_group_map[device_id]; }
+int AUTDGeometry::GroupIDForDeviceIdx(int device_idx) { return this->_group_map[device_idx]; }
 
 const Vector3 AUTDGeometry::position(int transducer_id) {
   const auto local_trans_id = transducer_id % NUM_TRANS_IN_UNIT;
-  auto device = this->device(transducer_id);
-  const auto pos = device->global_trans_positions.col(local_trans_id);
+  const auto pos = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->global_trans_positions.col(local_trans_id);
   return Vector3(pos.x(), pos.y(), pos.z());
 }
 
-const Vector3 AUTDGeometry::local_position(int device_id, Vector3 global_position) {
-  auto device = this->_devices.at(device_id);
-  auto local_origin = device->global_trans_positions.col(0);
-  auto x_dir = device->x_direction;
-  auto y_dir = device->y_direction;
-  auto z_dir = device->z_direction;
-  auto _global_position = Eigen::Vector3d(global_position.x(), global_position.y(), global_position.z());
-  auto rv = _global_position - local_origin;
+const Vector3 AUTDGeometry::local_position(int device_idx, Vector3 global_position) {
+  const Device* device = this->_devices.at(device_idx).get();
+  const auto local_origin = device->global_trans_positions.col(0);
+  const Eigen::Vector3d x_dir = device->x_direction;
+  const Eigen::Vector3d y_dir = device->y_direction;
+  const Eigen::Vector3d z_dir = device->z_direction;
+  const auto _global_position = Eigen::Vector3d(global_position.x(), global_position.y(), global_position.z());
+  const auto rv = _global_position - local_origin;
   return Vector3(rv.dot(x_dir), rv.dot(y_dir), rv.dot(z_dir));
 }
 
 const Vector3 AUTDGeometry::direction(int transducer_id) { return z_direction(transducer_id); }
 
 const Vector3 AUTDGeometry::x_direction(int transducer_id) {
-  const auto dir = this->_devices.at(this->deviceIdForTransIdx(transducer_id))->x_direction;
+  const Eigen::Vector3d dir = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->x_direction;
   return Vector3(dir.x(), dir.y(), dir.z());
 }
 
 const Vector3 AUTDGeometry::y_direction(int transducer_id) {
-  const auto dir = this->_devices.at(this->deviceIdForTransIdx(transducer_id))->y_direction;
+  const Eigen::Vector3d dir = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->y_direction;
   return Vector3(dir.x(), dir.y(), dir.z());
 }
 
 const Vector3 AUTDGeometry::z_direction(int transducer_id) {
-  const auto dir = this->_devices.at(this->deviceIdForTransIdx(transducer_id))->z_direction;
+  const Eigen::Vector3d dir = this->_devices.at(this->deviceIdxForTransIdx(transducer_id))->z_direction;
   return Vector3(dir.x(), dir.y(), dir.z());
 }
 
-const int AUTDGeometry::deviceIdForDeviceIdx(int device_idx) { return this->_devices.at(device_idx)->device_id; }
-
-const int AUTDGeometry::deviceIdForTransIdx(int transducer_id) { return this->device(transducer_id)->device_id; }
+const int AUTDGeometry::deviceIdxForTransIdx(int transducer_id) { return transducer_id / NUM_TRANS_IN_UNIT; }
 }  // namespace autd